+++ src/main/java/bebot/Deadline.java
@@ -1,4 +1,4 @@
-package duke;
+package bebot;
 
 import java.time.DateTimeException;
 import java.time.LocalDate;
