+++ src/main/java/bebot/Events.java
@@ -1,4 +1,4 @@
-package duke;
+package bebot;
 
 import java.time.DateTimeException;
 import java.time.LocalDate;
