+++ src/main/java/bebot/Parser.java
@@ -1,4 +1,4 @@
-package duke;
+package bebot;
 
 /**
  * Parser class deals with making sense of the user command
