+++ src/main/java/bebot/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package bebot;
 import java.util.Scanner;
 
 /**
