+++ src/test/java/bebot/TodoTest.java
@@ -1,6 +1,10 @@
-package duke;
+package bebot;
 
 import org.junit.jupiter.api.Test;
+
+import bebot.Storage;
+import bebot.Task;
+import bebot.TaskList;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
